pyproject.toml
[tool.poetry]
name = "volga-it"
version = "0.1.0"
description = ""
authors = ["shmookoff <shmookoff@gmail.com>"]
readme = "README.md"
packages = [{include = "volga_it"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.22"
sqlalchemy-mixins = "^2.0.5"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
uvicorn = "^0.23.2"
alembic = "^1.12.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.12"}
fastapi-sqlalchemy = "^0.2.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.6.1"
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
